update Scaling function

+++ Transformation.h
@@ -8,9 +8,9 @@
 #include "windows.h"
 
 
-void TranslationLineDDA(HDC hdc, int x1, int y1, int x2, int y2, int tx, int ty, COLORREF c);
+void TranslationLineDDA(HDC hdc, int x1, int y1, int x2, int y2, double tx, double ty, COLORREF c);
 
-void ScalingLineDDA(HDC hdc, double x1, double y1, double x2, double y2, int ScX, int ScY, COLORREF c);
+void ScalingLineDDA(HDC hdc, double x1, double y1, double x2, double y2, double ScX, double ScY, COLORREF c);
 
 
 #endif //GRAPHICS_PROJECT_TRANSFORMATION_H
